* less significant figures shown for colorschemeadd\n * more significant figures for g\n * g is now a float
git-svn-id: svn://gakman.com/gravit/trunk@299 60970b36-f649-e149-802f-859a9d0a703e

+++ gravit.h
@@ -343,7 +343,7 @@ typedef struct state_s {
 	int historyFrames;
 	int historyNFrame;
 
-	int gbase;
+	float gbase;
 	float g;
 
 	int frameCompression;
